feat(ErrorState): create generic error state component
