+++ common/error-state/styles.scss
@@ -1,0 +1,16 @@
+.ErrorState {
+  height: 100%;
+  text-align: center;
+}
+
+.ErrorState-cardBody {
+  height: 100%;
+  box-sizing: border-box;
+  display: flex;
+  flex-direction: column;
+  justify-content: center;
+}
+
+.ErrorState-headingText {
+  color: darkgrey;
+}
